+++ app/test/widget_test.dart
@@ -19,10 +19,10 @@ void main() {
     final homeRepo = MockHomeRepositoryImpl();
     final appLocalDataSource = AppLocalDataSource(packageInfo: packageInfo);
     final appRepository = AppRepositoryImpl(appLocalDataSource);
-    final themeRepository = ThemeRepositoryImpl(ThemeLocalDataSource(storage));
+    final themeRepository = ThemeRepositoryImpl(ThemeLocalDataSourceImpl(storage));
     final authRepository = AuthRepositoryImpl(
-      localDataSource: AuthLocalDataSource(storage),
-      remoteDataSource: AuthRemoteDataSource(
+      localDataSource: AuthLocalDataSourceImpl(storage),
+      remoteDataSource: AuthRemoteDataSourceImpl(
         client: remoteClient,
         storage: storage,
         soccialAuth: MockSccialAuth(),
